requirements.txt
python_Levenshtein==0.27.1
scikit_learn==1.4.2
setuptools==75.8.0
unidecode==1.3.8
